% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{actionButtonQW}
\alias{actionButtonQW}
\title{Create a themed actionButton}
\usage{
actionButtonQW(inputId, label, icon = NULL, style = "material-flat",
  color = "default", size = "xs", block = FALSE, no_outline = TRUE,
  dropdown = FALSE, tooltip = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link-usually a text label, but you
could also use any other HTML, like an image.}

\item{icon}{An optional \code{\link[shiny]{icon}} to appear on the button.}

\item{style}{Style of the button, to choose between \code{simple},
\code{bordered}, \code{minimal}, \code{stretch}, \code{jelly}, \code{gradient},
\code{fill}, \code{material-circle}, \code{material-flat}, \code{pill},
\code{float}, \code{unite}.}

\item{color}{Color of the button: \code{default}, \code{primary},
\code{warning}, \code{danger}, \code{success}, \code{royal}.}

\item{size}{Size of the button: \code{xs}, \code{sm}, \code{md}, \code{lg}.}

\item{block}{Logical, full width button.}

\item{no_outline}{Logical, don't show outline when navigating with keyboard/
interact using mouse or touch.}

\item{dropdown}{Logical, button displayed in dropdown.}

\item{tooltip}{Text to appear as a tooltip on the button.}
}
\description{
Create a themed actionButton
}
