% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_factors.R
\name{all_effects}
\alias{all_effects}
\title{All possible effects for a simple linear model between factors}
\usage{
all_effects(factors)
}
\arguments{
\item{factors}{A character vector containing the factor names.}
}
\description{
All possible effects for a simple linear model between factors
}
