% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checked_text_input.R
\name{checked_text_input}
\alias{checked_text_input}
\alias{checked_text_input_ui}
\title{Checked text input}
\usage{
checked_text_input(input, output, session, .data, .values, .parent,
  .label = NULL, .value = "", .allow_reset = TRUE, .reset = NULL,
  .update_value = NULL, .error_controller = NULL,
  .inline_block = FALSE)

checked_text_input_ui(id)
}
\arguments{
\item{input, output, session}{Called by \code{\link[shiny]{callModule}}.}

\item{.data}{The \code{.data} object.}

\item{.values}{The \code{.values} list.}

\item{.parent}{The parent \code{\link{Node}} object.}

\item{.label}{The contents of the button or link-usually a text label, but
you could also use any other HTML, like an image.}

\item{.value}{Function without an argument (e.g
\code{\link[shiny:reactive]{reactive}}) or vector containing the initial
value of the text input.}

\item{.allow_reset}{If \code{\link[base:logical]{TRUE}}, reset to the initial
value via an \code{\link[shiny:actionButton]} is enabled.}

\item{.reset}{\code{\link[shiny:reactive]{Reactive}}. See 'Details'.}

\item{.update_value}{\code{\link[shiny:reactive]{Reactive}}. See 'Details'.}

\item{.error_controller}{An \code{\link{ErrorController}} or
\code{\link{ErrorControllerList}}}

\item{id}{The module's id.}
}
\value{
\code{checked_text_input} returns a list of \code{\link[shiny:reactive]{reactives}}:

\tabular{ll}{
  \code{name} \tab Current value of the text input. \cr
  \code{null_name} \tab If an error occured \code{\link[base:NULL]{NULL}}, else
  the current value of the text input. \cr
  \code{error} \tab If an error occured \code{\link[base:logical]{TRUE}},
  else \code{\link[base:logical]{FALSE}}.
}
}
\description{
textInput whose value is checked for being useable as an object name
}
\details{
When the value of the reactive passed to \code{.reset} changes, the textInput
is reset to the last updated value. When the value of the reactive passed to
\code{.update_value} changes the current value of the text input gets the
last updated value.
}
