% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_chart_functions.R
\name{control_chart_function}
\alias{control_chart_function}
\alias{A1}
\alias{A2}
\alias{A3}
\alias{B3}
\alias{B4}
\alias{c4}
\alias{d2}
\alias{D3}
\alias{D4}
\title{Control Chart Functions}
\usage{
A1(n)

A2(n)

A3(n)

B3(n)

B4(n)

c4(n)

d2(n)

D3(n)

D4(n)
}
\arguments{
\item{n}{Sample size.}
}
\description{
\code{A1}, \code{A2}, \code{d2}, \code{D3} and \code{D4} currently use
hard-coded values for sample size \code{n = 5}.
}
