% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_line_value.R
\name{control_line_value}
\alias{control_line_value}
\title{Compute control line values}
\usage{
control_line_value(data, quantile = c(-3, 0, 3), type = c("xbar_R",
  "xbar_s", "R", "s", "p"))
}
\arguments{
\item{data}{A \code{\link[tibble]{tibble}} with numeric columns \code{sample} and
\code{value}. All samples should be of the same size. Else the sample size is
determined by calculating the mean sample size.}

\item{quantile}{A numeric vector representing the quantiles for which the
control line values are calculated. If \code{type \%in\% c("xbar_R", "R")}
quantiles must be in \code{c(-3, 0, 3)}.}

\item{type}{Control chart's type.}
}
\description{
Compute control line values
}
