% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlChartControlLine.R
\docType{data}
\name{ControlChartControlLine}
\alias{ControlChartControlLine}
\title{ControlChartControlLine}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
ControlChartControlLine
}
\description{
ControlChartControlLine
}
\keyword{datasets}
