% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlChartPhase.R
\name{ControlChartPhase}
\alias{ControlChartPhase}
\title{ControlChartPhase}
\description{
R6Class representing a control chart phase consisting of multiple
\code{\link[ControlChartSample]{ControlChartSamples}}. Multiple phases may
form a \code{\link{ControlChart}}.
}
\section{Usage}{

\preformatted{phase = ControlChartPhase$new(name, preliminary = FALSE)
}
}

\section{Methods}{

\describe{
  \item{\code{new(name, preliminary = FALSE)}}{Initialize the phase.
    \describe{
      \item{\code{name}}{\code{\link[base:character]{Character}}. The name
      of the phase.}
      \item{\code{preliminary}}{If \code{\link[base:logical]{TRUE}}, the
      phase is used to calculate the control line values of the parent
      control chart, otherwise not.}
    }
  }
  \item{\code{add_sample(sample)}}{Add a \code{\link{ControlChartSample}} to
  the phase.
    \describe{
      \item{\code{sample}}{A \code{\link{ControlChartSample}}.}
    }
  }
  \item{\code{get_id()}}{Get the phase's id.
  }
  \item{\code{get_name()}}{Get the phase's name.
  }
  \item{\code{get_sample(id)}}{Get the phase's sample with \code{id == id}.
  }
  \item{\code{get_sample_ids()}}{Get all sample ids as a character vector.
  }
  \item{\code{get_value()}}{Get a \code{\link[tibble:tibble]{tibble}} with
  columns \code{"sample"} and \code{"value"}.
  }
  \item{\code{is_preliminary()}}{Returns a \code{\link[base:logical]{logical}}
  indicating whether the phase is preliminary or not.
  }
  \item{\code{remove_sample(id)}}{Remove the sample with \code{id == id}.
  }
  \item{\code{set_name()}}{Set the phase's name.
  }
  \item{\code{set_preliminary(value)}}{Set whether the phase is preliminary
  or not.
    \describe{
      \item{\code{value}}{A \code{\link[base:logical]{logical}}.}
    }
  }
}
}

