% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlChartSample.R
\name{ControlChartSample}
\alias{ControlChartSample}
\title{ControlChartSample}
\description{
R6Class representing a control chart sample. Multiple samples may form a
\code{\link{ControlChartPhase}}.
}
\section{Usage}{

\preformatted{sample = ControlChartSample$new(value, name)
}
}

\section{Methods}{

\describe{
  \item{\code{new(value, name = character())}}{Initialize the sample.
    \describe{
      \item{\code{value}}{A \code{\link[base:numeric]{numeric}} vector.}
      \item{\code{name}}{\code{\link[base:character]{character}}.}
    }
  }
  \item{\code{get_id()}}{Get the sample's id.
  }
  \item{\code{get_length()}}{Get the sample's length.
  }
  \item{\code{get_name()}}{Get the sample's name.
  }
  \item{\code{get_value()}}{Get the sample's value.
  }
  \item{\code{set_name(name)}}{Set the sample's name.
  }
  \item{\code{set_value(value)}}{Set the sample's value.
    \tabular{ll}{
      \code{value} \tab A \code{\link[base:numeric]{numeric}} vector.
    }
  }
}
}

