% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_directory_lines.R
\name{count_directory_lines}
\alias{count_directory_lines}
\title{Count the lines of all scripts inside a directory}
\usage{
count_directory_lines(directory, recursive = TRUE, output = c("sum",
  "separated"), remove = c("none", "comments", "roxygen"))
}
\arguments{
\item{directory}{A path to the directory containing the R scripts.}

\item{recursive}{If \code{\link[base:logical]{TRUE}}, subdirectories are
recursively searched for R scripts.}

\item{output}{If \code{"sum"}, calculate the sum of all lines; if
\code{"separated"}, show number of lines for each individual script.}

\item{remove}{If \code{"comments"}, all lines
starting with "#" don't add to the count. If \code{"roxygen"}, all lines
starting with "#'" dont' add to the count.}
}
\description{
Count the lines of all scripts inside a directory.
}
\details{
If used with \code{comment = "roxygen"} the results are comparable to
\code{\link{count_package_lines}}.
}
