% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_function_lines.R
\name{count_function_lines}
\alias{count_function_lines}
\title{Count lines of a function definition}
\usage{
count_function_lines(fun)
}
\arguments{
\item{fun}{Function name.}
}
\description{
Count lines of a function definition
}
