% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_lines.R
\name{count_lines}
\alias{count_lines}
\title{Count the lines of all scripts inside a directory}
\usage{
count_lines(directory, recursive = TRUE, output = c("sum",
  "separated"))
}
\arguments{
\item{directory}{A path to the directory containing the R scripts.}

\item{recursive}{If \code{\link[base:logical]{TRUE}}, subdirectories are
recursively searched for R scripts.}

\item{output}{If \code{"sum"}, calculate the sum of all lines; if
\code{"separated"}, show number of lines for each individual script.}
}
\description{
Count the lines of all scripts inside a directory.
}
