% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_package_lines.R
\name{count_package_lines}
\alias{count_package_lines}
\title{Count the lines of all functions in a loaded package}
\usage{
count_package_lines(package, output = c("sum", "separated"))
}
\arguments{
\item{package}{Package name.}

\item{Output}{If \code{"sum"}, calculate the sum of all lines; if
\code{"separated"}, show number of lines for each individual package
function.}
}
\description{
Count the lines of all functions in a loaded package, which is NOT equal to
the number of lines of the scripts defining the functions due to the absence
of the roxygen documentation. Furthermore only functions are considered and
not for example R6 objects.
}
