% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selector_column_dropdown.R
\name{data_selector_column_dropdown_ui}
\alias{data_selector_column_dropdown_ui}
\alias{data_selector_column_dropdown}
\title{Dropdown for column actions}
\usage{
data_selector_column_dropdown_ui(id)

data_selector_column_dropdown(input, output, session, .data, .values,
  .parent, group_name, dataset_name, column_name)
}
\arguments{
\item{id}{The module's id}

\item{input, output, session}{Called by \code{\link[shiny:callModule]{callModule}}.}

\item{.data}{The \code{.data} object.}

\item{.values}{The \code{.values} list.}

\item{.parent}{The parent \code{\link{Node}} object.}

\item{group_name, dataset_name, column_name}{\code{\link[shiny:reactive]{Reactives}}
containing the group, dataset and column selected by the user.}
}
\description{
Dropdown for column actions
}
