% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSelector.R
\name{data_selector_output}
\alias{data_selector_output}
\title{data_selector_output}
\usage{
data_selector_output(id, element_id)
}
\arguments{
\item{id}{id of the corresponding \code{\link{DataSelector}}.}

\item{element_id}{id of the element to display}
}
\description{
data_selector_output
}
