% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSelector.R
\docType{data}
\name{DataSelector}
\alias{DataSelector}
\alias{data_selector_output}
\title{DataSelector}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataSelector

data_selector_output(id, element_id)
}
\arguments{
\item{id}{id of the corresponding \code{\link{DataSelector}}.}

\item{element_id}{id of the element to display}
}
\description{
\code{\link[R6:R6Class]{R6Class}} for generating customised selections of
groups, datasets and columns. After initialing the DataSelector an arbitrary
amout of elements may be added in the \code{\link[shiny:shiny-package]{shiny}}
server function. After adding all elements the DataSelector gets activated
by invoking the method \code{this$callModule()}. The display of the elements
is handled by \code{data_selector_output}. For each element of a
\code{\link{DataSelector}} \code{data_selector_output} has to be called
either in the \code{ui} function or inside of
\code{\link[shiny:renderUI]{renderUI}} to display the UI of the corresponding
element.
}
\section{Usage}{

\preformatted{DataSelector$new(
  id = "data_selector"
)$Group(
  id = "group",
  label = "Select group"
)$Dataset(
  id = "dataset",
  group_id = "group",
  label = "Select dataset
)$Column(
  id = "column_1",
  group_id = "group",
  dataset_id = "dataset",
  label = "Select column 1",
  setdiff_id = "column_2"
)$Column(
  id = "column_2",
  group_id = "group",
  dataset_id = "dataset",
  label = "Select column 2",
  setdiff_id = "column_1"
)$callModule(
  .data = data,
  .values = values,
  .parent = parent
)
}
}

\section{Methods}{

\describe{
  \item{\code{new(id)}}{Initialize the data_selector.
    \tabular{ll}{
      \code{id} \tab The id of the data_selector.
    }
  }
  \item{\code{callModule(.data, .values, .parent)}}{Activate the data_selector
    inside the server function. Before calling this function NO ui will be
    displayed.
    \tabular{ll}{
      \code{.data} \tab \code{\link{DataStorage}}. \cr
      \code{.values} \tab List of \code{\link[shiny:reactive]{reactive}}-like objects
      as specified by XXX. \cr
      \code{.parent} \tab \code{\link{Node}}.
    }
  }
  \item{\code{Column(id, group_id, dataset_id, label, extended = TRUE,
    setdiff_id = NULL, multiple = FALSE)}}{Add a
    \code{\link[shiny:selectInput]{selectInput}} for selecting a column
    of a dataset.
    \tabular{ll}{
      \code{id} \tab \code{\link[base:character]{Character}}. The id of the
      column element. \cr
      \code{group_id} \tab \code{\link[base:character]{Character}}. The id
      of the parent group element. \cr
      \code{dataset_id} \tab \code{\link[base:character]{Character}}. The id
      of the parent dataset element. \cr
      \code{label} \tab The label of the
        \code{\link[shiny:selectInput]{selectInput}}. \cr
      \code{extended} \tab \code{\link[base:logical]{Logical}}. If
        \code{\link[base:logical]{TRUE}}, the label of the
        \code{\link[shiny:selectInput]{selectInput}} is extended by a
        \code{\link{data_selector_column_dropdown}}. \cr
      \code{setdiff_id} \tab \code{\link[base:character]{Character}} vector.
        The  value of the \code{\link[shiny:selectInput]{selectInput}} is
        compared to the values of all column elements with
        \code{id \%in\% setdiff_id}. If one of these elements has the same
        value, this element's \code{\link[shiny:selectInput]{selectInput}}
        gets updated to the first free column.
        If \code{multiple =} \code{\link[base:logical]{TRUE}}, no comparison
        is done.
        \cr
      \code{multiple} \tab \code{\link[base:logical]{Logical}}. If
        \code{\link[base:logical]{TRUE}}, multiple columns can be selected.
    }
  }
  \item{\code{Dataset(id, group_id, label, extended = TRUE)}}{Add a
    \code{\link[shiny:selectInput]{selectInput}} for selecting a dataset
    of a group.
    \tabular{ll}{
      \code{id} \tab The id of the dataset element. \cr
      \code{group_id} \tab The id of the parent group element. \cr
      \code{label} \tab The label of the
        \code{\link[shiny:selectInput]{selectInput}}. \cr
      \code{extended} \tab \code{\link[base:logical]{Logical}}. If
        \code{\link[base:logical]{TRUE}}, the label of the
        \code{\link[shiny:selectInput]{selectInput}} is extended by a
        \code{\link{data_selector_dataset_dropdown}}.
    }
  }
  \item{\code{Group(id, label, extended = TRUE)}}{Add a
    \code{\link[shiny:selectInput]{selectInput}} for selecting a group.
    \tabular{ll}{
      \code{id} \tab The id of the group element. \cr
      \code{label} \tab The label of the
        \code{\link[shiny:selectInput]{selectInput}}. \cr
      \code{extended} \tab \code{\link[base:logical]{Logical}}. If
        \code{\link[base:logical]{TRUE}}, the label of the
        \code{\link[shiny:selectInput]{selectInput}} is extended by a
        \code{\link{data_selector_column_dropdown}}.
    }
  }
}
}

\keyword{datasets}
