% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe_plots.R
\name{doe_contour_plot}
\alias{doe_contour_plot}
\title{Contour Plot}
\usage{
doe_contour_plot(fac_design, factor_1, factor_2, interactions = FALSE)
}
\arguments{
\item{fac_design}{A \code{\link{FacDesign}} object.}

\item{factor_1, factor2}{A \code{\link{FacDesignFactor}} object in
\code{fac_design$get_factor_storage()}.}

\item{interactions}{If \code{\link[base:logical]{TRUE}}, interactions between
\code{factor_1} and \code{factor_2} are considered.}
}
\description{
Visualize the prediction of a full factorial design for two factors with
a two-dimensional contour plot.
}
