% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe_plots.R
\name{doe_interaction_plot}
\alias{doe_interaction_plot}
\title{Interaction Plot}
\usage{
doe_interaction_plot(fac_design, factor_1, factor_2)
}
\arguments{
\item{fac_design}{A \code{\link{FacDesign}} object.}

\item{factor_1, factor2}{A \code{FacDesignFactor} object in
\code{fac_design$get_factor_storage()}.}
}
\description{
Visualize the interaction between two factors of a full factorial design.
}
