% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe_plots.R
\name{doe_pareto_plot}
\alias{doe_pareto_plot}
\title{Pareto Plot}
\usage{
doe_pareto_plot(fac_design, lm, alpha = 0.05,
  title = "Standardisierte Haupt- und Wechselwirkungen")
}
\arguments{
\item{lm}{A linear model generated by \code{\link[stats]{lm}}.}

\item{alpha}{Level of significance.}

\item{title}{The title of the plot.}

\item{A}{\code{\link{FacDesign}} object.}
}
\description{
Visualize the significance of effects of a linear model. Currently this methods
needs a \code{\link{FacDesign}} object as input in addition to the linear model.
}
