% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorControllerList.R
\name{ErrorControllerList}
\alias{ErrorControllerList}
\title{ErrorControllerList}
\description{
Reactive R6Class that returns HTML according to a test function. Modules like
\code{\link{checked_text_input}} or \code{\link{observed_vector_input}}
provide a argument accepting an \code{\link{ErrorController}} or
ErrorControllerList that expands the module's default tests.
}
\section{Usage}{

\preformatted{
error_controller <- shiny::isolate({
  ErrorControllerList$new()$add_error_controller(
    ErrorController$new(
      fun = function(x) purrr::is_null(x),
      ui_true = "x is NULL",
      ui_false = "x is not NULL",
      error = TRUE
    )
  )})
}
}

\section{Methods}{

\describe{
  \item{new()}{Initialize the error controller list.
  }
  \item{add_error_controller(error_controller)}{Add an \code{\link{ErrorController}}
    to the error controller list.
  }
  \item{get_ui()}{Get the UI of all \code{\link{ErrorController}} as a list
    ordered by the priority of the respective \code{\link{ErrorController}}.
  }
  \item{has_error()}{Returns \code{\link[base:logical]{TRUE}} if any of
  the \code{\link{ErrorController}} has an error.
  }
  \item{set_value(value)}{Sets the value of all \code{\link{ErrorController}}.
  }
}
}

