% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_list.R
\name{expand_list}
\alias{expand_list}
\title{Easy way to add an empty list to a list}
\usage{
expand_list(.list, position)
}
\arguments{
\item{.list}{A list.}

\item{position}{A character vector.}
}
\description{
Add an empty list to a list at a position specified by a character vector,
which is interpreted recursively. See \code{Examples}.
}
\details{
\code{\link[base:Extract]{`[[`}} allows indexing a list with a vector, but only
if the indices are already present in the list. \code{expand_list} implements
an easy way to equip a list with the needed indices.
}
\examples{
l <- list()
# Add empty list "a" on top level
l <- expand_list(l, "a")
# Add list "b" on top level with empty list "c" inside it
l <- expand_list(l, c("b", "c"))
# Add empty list "d" to the previously defined list "b"
l <- expand_list(l, c("b", "d"))

}
