% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacDesign.R
\name{FacDesign}
\alias{FacDesign}
\title{FacDesign}
\description{
R6Class representing a factorial design.
}
\section{Usage}{

\preformatted{fac_design <- FacDesign$new(name, k)
}
}

\section{Methods}{

\describe{
  \item{\code{new(name, k, rep = 1, center = 0, fac_names = LETTERS[seq_len(k)],
  response_name = "response", index_name = "index", response = NULL)}}{
  Initialize the factorial design.
    \tabular{ll}{
      \code{name} \tab \code{\link[base:character]{Character}}. The factorial
      design's name \cr
      \code{k} \tab \code{\link[base:integer]{Integer}}. Number of variable
      factors. \cr
      \code{center} \tab \code{\link[base:integer]{Integer}}. Number of
      center points. \cr
      \code{fac_names} \tab A \code{\link[base:character]{character}} vector
      containing the names of the variable factors. \cr
      \code{response_name} \tab \code{\link[base:character]{Character}}. Name
      of the response variable. \cr
      \code{index_name} \tab \code{\link[base:character]{Character}}. Name of
      the index variable. \cr
      \code{response} \tab A \code{\link[base:numeric]{numeric}} vector
      containing the values of the response variable for the standardised
      factorial design. If \code{response = NULL}, you have to add a response
      later via \code{this$add_response()}.
    }
  }
  \item{\code{add_response(response)}}{Add or update the response of the
  factorial design.
    \tabular{ll}{
      \code{response} \tab A \code{\link[base:numeric]{numeric}} vector
      containing the values of the response variable for the standardised
      factorial design.
    }
  }
  \item{\code{get_name()}}{Get the factorial design's name.
  }
  \item{\code{get_names(fac = FALSE, index = FALSE, name = FALSE, response = FALSE)}}{
  Get a list with the following entries:
    \tabular{ll}{
      fac \tab The variable factors' names. \cr
      index \tab The index' name. \cr
      name \tab The factorial design's name. \cr
      response \tab The factorial design response's name.
    }
   The arguments decide wheter a name is part of the list or not.
  }
  \item{\code{get_table()}}{
  }
}
}

