% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_factors.R
\name{factors_to_formula_factory}
\alias{factors_to_formula_factory}
\title{Factory generating a formula for multiple linear regression}
\usage{
factors_to_formula_factory(factor_indices, interaction_level)
}
\arguments{
\item{factor_indices}{Integer vector.}

\item{interaction_level}{Maximum length of interaction between factors.}
}
\description{
Get a function representing the right side of a formula with interactions
between factors up to the specified interaction level. Call this function
with the actual factors to get an up-to-date formula.
}
