% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_factors.R
\name{factors_to_formula_text}
\alias{factors_to_formula_text}
\title{Generate formula for multiple linear regression}
\usage{
factors_to_formula_text(factors, interaction_level = 0)
}
\arguments{
\item{factors}{Factor names.}

\item{interaction_level}{Maximum length of interaction between factors.}
}
\description{
Get a character representing the right side of a formula with interactions
between factors up to the specified interaction_level.
}
