% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fallback}
\alias{fallback}
\title{Handle NULL with fallback value}
\usage{
fallback(x, y)
}
\arguments{
\item{x}{Possible NULL value}

\item{y}{Fallback value}
}
\description{
This function is especially useful in \code{\link[shiny]{renderUI}} when an
input value has not yet been initialised
}
