% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer_node_functions.R
\name{get_ancestor_node}
\alias{get_ancestor_node}
\title{Get ancestor node of a node}
\usage{
get_ancestor_node(node, generation)
}
\arguments{
\item{node}{A \code{\link{ExplorerNode}}.}

\item{generation}{An \code{\link[base:integer]{integer}} indicating the number
of generations to go back in the node tree. Negative integers and 0 will result
in returning \code{node}.

If generation is greater than the number of the node's ancestors, NULL is
returned.}
}
\description{
Get the ancestor node of a node which is \code{generation} generations away
from \code{node}.
}
