% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer_node_functions.R
\name{get_node_distance}
\alias{get_node_distance}
\title{Get distance of a node to the root node}
\usage{
get_node_distance(node)
}
\arguments{
\item{node}{A \code{\link{ExplorerNode}}.

If \code{node} is the root node, \code{0} is returned.}
}
\description{
Get the distance of a node to the root node.
}
