% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_static.R
\name{handle_static_counter}
\alias{handle_static_counter}
\title{Handle assignment of a static counter for R6 classes}
\usage{
handle_static_counter(static, name = "counter", initial_value = 1)
}
\arguments{
\item{static}{\code{\link[base:environment]{Environment}}, which holds the
counter.}

\item{name}{Name of the variable representing the counter.}

\item{initial_value}{Initial value of the variable representing the counter.}
}
\description{
This is a convenience function, checking if a static counter is already
initialised. If not, it initialises it with \code{initial_value}, otherwise
it adds one to it.
}
