% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTest.R
\name{HypothesisTest}
\alias{HypothesisTest}
\title{Hypothesis Test}
\description{
\code{\link[R6:R6Class]{R6Class}} representing a hypothesis test consisting
of \code{\link{HypothesisTestSample}}s and \code{\link{NullHypothesis}}.
}
\section{Usage}{

\preformatted{hypothesis_test = HypothesisTest$new()

hypothesis_test$add_sample(HypothesisTestSample$new())

hypothesis_test$add_null_hypothesis(NullHypothesis$new())
}
}

\section{Methods}{

\describe{
  \item{\code{new()}}{Initialize the hypothesis test.
  }
  \item{\code{add_null_hypothesis(null_hypothesis)}}{Add a 
  \code{\link{NullHypothesis}} to the hypothesis test.
    \tabular{ll}{
      \code{hypothesis_test} \tab A \code{\link{NullHypothesis}} object.
    }
  }
  \item{\code{add_sample(sample)}}{Add a \code{\link{HypothesisTestSample}} 
  to the hypothesis test.
    \tabular{ll}{
      \code{sample} \tab A \code{\link{HypothesisTestSample}} object.
    }
  }
  \item{\code{get_null_hypothesis(id)}}{Get the hypothesis test's null
  hypothesis with \code{id == id}.
  }
  \item{\code{get_null_hypothesis_ids()}}{Get all null hypothesis ids as a
  \code{\link[base:character]{character}} vector.
  }
  \item{\code{get_sample(id)}}{Get the hypothesis tests's sample with
  \code{id == id}.
  }
  \item{\code{get_sample_ids()}}{Get all sample ids as a 
  \code{\link[base:character]{character}} vector.
  }
  \item{\code{get_id()}}{Get the hypothesis test's id.
  }
  \item{\code{get_name()}}{Get the hypothesis test's name.
  }
  \item{\code{remove_null_hypothesis(id)}}{Remove the null hypothesis with
  \code{id == id}.
  }
  \item{\code{remove_sample(id)}}{Remove the sample with \code{id == id}.
  }
  \item{\code{set_name(name)}}{Set the hypothesis test's name.
  }
}
}

