% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTestSample.R
\name{HypothesisTestSample}
\alias{HypothesisTestSample}
\title{HypothesisTestSample}
\description{
\code{\link[R6:R6Class]{R6Class}} representing a sample of a
\code{HypothesisTest}
}
\section{Usage}{

\preformatted{sample <- HypothesisTestSample$new()
}
}

\section{Methods}{

\describe{
  \item{\code{new(name = "", value = numeric())}}{Initialize a new sample
  object.
    \tabular{ll}{
      \code{name} \tab \code{\link[base:character]{Character}}. The sample's
      name. \cr
      \code{value} \tab A vector (not necessarily numeric) representing the
      sample's value.
    }
  }
  \item{\code{get_id()}}{Get the sample's id.
  }
  \item{\code{get_name()}}{Get the sample's name.
  }
  \item{\code{get_value()}}{Get the sample's value.
  }
  \item{\code{set_name(name)}}{Set the sample's name with \code{name}.
  }
  \item{\code{set_value(value)}}{Set the sample's value with \code{value}.
  }
}
}

