% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{label_lang}
\alias{label_lang}
\title{Create label in multiple languages}
\usage{
label_lang(...)
}
\arguments{
\item{...}{Named characters.}
}
\description{
This function returns text in the language specified by the \code{.language}
global variable. If this variable is not defined the first language will be
returned.
}
