% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{label_lang_list}
\alias{label_lang_list}
\title{Create choices for shiny inputs in multiple languages}
\usage{
label_lang_list(..., value)
}
\arguments{
\item{...}{Named character vectors.}

\item{value}{Character vector with unique values that will be returned by
input$<inputId>}
}
\description{
Some shiny inputs like \code{\link[shiny]{selectInput}} need a named list
for their choices.
}
