% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{length_attr<-}
\alias{length_attr<-}
\title{Guarantee that attribute has the same length as the object.}
\usage{
length_attr(x, name) <- value
}
\arguments{
\item{x}{An \code{R} object.}

\item{name}{The attribute's name.}

\item{value}{A vector of the same length as x, or \code{NULL}.}
}
\description{
Guarantee that attribute has the same length as the object.
}
