% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{maprange}
\alias{maprange}
\title{Map value from a previous to a new range}
\usage{
maprange(x, pmin, pmax, nmin, nmax)
}
\arguments{
\item{x}{Numeric value.}

\item{pmin}{Previous minimum.}

\item{pmax}{Previous maximum.}

\item{nmin}{New minimum.}

\item{nmax}{New maximum.}
}
\description{
Map value from a previous to a new range
}
