% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numeric_to_vector_text}
\alias{numeric_to_vector_text}
\title{Convert numeric to string}
\usage{
numeric_to_vector_text(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\description{
Convert a numeric vector to a string in which each number is seperated by
\code{,}.
}
