% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{Object}
\alias{Object}
\title{Object}
\description{
\code{\link[R6]{R6Class}} representing the most primitive object having only
an id and a name.
}
\section{Methods}{

\describe{
  \item{\code{new(name = "")}}{Initialize a new object with name \code{name}.
  }
  \item{\code{get_id()}}{Get the object's id.
  }
  \item{\code{get_name()}}{Get the object's name.
  }
  \item{\code{set_name(name)}}{Set the object's name. Only applicable if
    \code{name} is not a \code{\link[shiny]{reactive}}.
  }
}
}

