% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_list.R
\name{object_list}
\alias{object_list}
\title{Get name and type of objects from all the scripts of a directory}
\usage{
object_list(directory, recursive, chdir = FALSE, all.names = FALSE)
}
\arguments{
\item{directory}{A path to the directory containing the R scripts.}

\item{recursive}{If \code{\link[base:logical]{TRUE}}, subdirectories are
recursively searched for R scripts.}

\item{chdir}{If \code{\link[base:logical]{TRUE}}, the \code{R} working
directory is temporarily changed to the directory containing the R script.}

\item{all.names}{Logical. If \code{\link[base:logical]{TRUE}} all object
names are returned. If \code{\link[base:logical]{FALSE}}, names which begin
with a . are omitted.}
}
\value{
A list of the objects. The path from the root of the list to an object
corresponds to the path of the R script.
}
\description{
Get name and type of all objects that are assigned on the top level of the
R scripts in a directory.
}
