% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_names.R
\name{object_names}
\alias{object_names}
\title{Get the names of all objects assigned in an R script}
\usage{
object_names(path, chdir = FALSE, all.names = FALSE)
}
\arguments{
\item{path}{Path to the R script.}

\item{chdir}{If \code{\link[base:logical]{TRUE}}, the \code{R} working
directory is temporarily changed to the directory containing the R script.}

\item{all.names}{Logical. If \code{\link[base:logical]{TRUE}} all object
names are returned. If \code{\link[base:logical]{FALSE}}, names which begin
with a . are omitted.}
}
\description{
Get the names of all objects assigned in an R script.
}
