% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectStorage.R
\name{ObjectStorage}
\alias{ObjectStorage}
\title{ObjectStorage}
\description{
R6Class for storing similar R6 objects.
}
\section{Usage}{

\preformatted{storage = ObjectStorage$new(allowed_classes = NULL)

storage$add_object(object)

storage$get_object(name)
}
}

\section{Methods}{

\describe{
  \item{\code{new(allowed_classes = NULL)}}{Initialize the storage.
    \describe{
      \item{\code{allowed_classes}}{Character vector. \code{class(object)} has to
      return at least one of these classes for being added to the storage.}
    }
  }
  \item{\code{add_object(object)}}{Add an object to the storage.
    \describe{
      \item{\code{object}}{R6 object with public method \code{get_name()}.}
    }
  }
  \item{\code{get_names()}}{Get the names of the stored objects as a
    character vector.
  }
  \item{\code{get_object(name, lazy = FALSE)}}{Get an object from the storage
    with \code{object$get_name() == name}.
    \describe{
      \item{\code{name}}{Name of an R6 object.}
      \item{\code{lazy}}{If \code{\link[base:logical]{TRUE}}, allow a name
        which is not present in the names of the storage for a short time.
      }
    }
  }
  \item{\code{get_objects(names)}}{Get a list of objects from the storage
    with \code{object$get_name() \%in\% names}.
    \describe{
      \item{\code{names}}{Character vector. Each element has to be a name of
        an object in the storage.
      }
    }
  }
}
}

