% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_vector_input.R
\name{observed_vector_input}
\alias{observed_vector_input}
\alias{observed_vector_input_ui}
\title{Observed vector input}
\usage{
observed_vector_input(input, output, session, .values, .parent,
  .label = NULL, .value = "", .error_controller_text = NULL,
  .error_controller_value = NULL)

observed_vector_input_ui(id)
}
\arguments{
\item{input, output, session}{Called by \code{\link[shiny:callModule]{callModule}}.}

\item{.values}{The \code{.values} list.}

\item{.parent}{The parent \code{\link{Node}} object.}

\item{.label}{The contents of the button or link-usually a text label, but
you could also use any other HTML, like an image.}

\item{.value}{Function without an argument (e.g
\code{\link[shiny:reactive]{reactive}}) or vector containing the initial
value of the text input.}

\item{.error_controller_text}{\code{\link{ErrorController}} or
\code{\link{ErrorControllerList}} for controlling the text of the
\code{\link[shiny:textInput]{textInput}}.}

\item{.error_controller_value}{\code{\link{ErrorController}} or
\code{\link{ErrorControllerList}} for controlling the module's value
(\code{\link[base:numeric]{numeric}} vector).}

\item{id}{The module's id.}
}
\description{
Observed vector input
}
