% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{package_functions}
\alias{package_functions}
\title{Get the name of all functions in a loaded package}
\usage{
package_functions(package)
}
\arguments{
\item{package}{Package name.}
}
\description{
Get the name of all functions in a loaded package
}
