% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_member.R
\name{reactive_member}
\alias{reactive_member}
\title{Reactive member}
\usage{
reactive_member(value)
}
\arguments{
\item{value}{If \code{\link[shiny:reactive]{reactive}}, this reactive is
returned, else a \code{\link[shiny:reactiveVal]{reactiveVal}} with value as
its initial value is returned.}
}
\description{
Generate a \code{\link[R6:R6Class]{R6Class}}-member with reactive behaviour.
}
