% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{searchInputQW}
\alias{searchInputQW}
\title{Search Input}
\usage{
searchInputQW(inputId, label = NULL, value = "", placeholder = NULL,
  btnSearch = NULL, btnReset = NULL, resetValue = "", width = NULL,
  disable_text = FALSE, disable_search = FALSE,
  disable_reset = FALSE)
}
\description{
Search Input
}
