% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_input.R
\name{significance_input_ui}
\alias{significance_input_ui}
\alias{significance_input}
\title{Input module for level of significance}
\usage{
significance_input_ui(id)

significance_input(input, output, session, .data, .values, .parent)
}
\arguments{
\item{id}{The module's id.}

\item{input, output, session}{Called by \code{\link[shiny:callModule]{callModule}}.}

\item{.data}{The \code{.data} object.}

\item{.values}{The \code{.values} list.}

\item{.parent}{The parent \code{\link{Node}} object.}
}
\value{
\code{significance_input} returns a list of \code{\link[shiny:reactive]{reactives}}:
\tabular{ll}{
  \code{alpha} \tab The level of significance, which is one minus the level
  of confidence.
}
}
\description{
User may input the level of significance or
confidence, choose between a \code{\link[shiny:sliderInput]{sliderInput}} or
a \code{\link[shiny:numericInput]{numericInput}} and gets a quick choice with
three \code{\link{actionButtonQW}}.
}
