% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_directory.R
\name{source_directory}
\alias{source_directory}
\title{Source files recursively}
\usage{
source_directory(path, pattern = ".*[.](r|R|s|S|q)([.](lnk|LNK))*$",
  recursive = TRUE, envir = parent.frame(), onError = c("error",
  "warning", "skip"), modifiedOnly = TRUE, ..., verbose = FALSE)
}
\arguments{
\item{path}{A path to a directory to be sourced.}

\item{pattern}{A regular expression file name pattern to identify
       source code files.}

\item{recursive}{If \code{\link[base:logical]{TRUE}}, subdirectories are recursively sourced
       first, otherwise not.}

\item{envir}{An \code{\link[base]{environment}} in which the code should be evaluated.}

\item{onError}{If an error occurs, the error may stop the job,
       give a warning, or silently be skipped.}

\item{modifiedOnly}{If \code{\link[base:logical]{TRUE}}, only files that are modified since the
       last time they were sourced are sourced, otherwise regardless.}

\item{...}{Additional arguments passed to \code{\link{sourceTo}}().}

\item{verbose}{A \code{\link[base]{logical}} or a \code{\link{Verbose}} object.}
}
\description{
Modification of \code{\link[R.utils]{sourceDirectory}} which repairs the
behaviour of the ... argument, so that \code{chdir} and \code{local} can be
passed to \code{\link[R.utils]{sourceTo}}.
}
