% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_factors.R
\name{two_factor_formula}
\alias{two_factor_formula}
\title{Get a string representing a simple linear model two-factor explanation of a
response variable.}
\usage{
two_factor_formula(response_name, factor_1, factor_2,
  interactions = FALSE)
}
\arguments{
\item{response_name}{Name of the response variable.}

\item{factor_1, factor_2}{A \code{\link{FacDesignFactor}} object.}

\item{interactions}{If \code{\link[base:logical]{TRUE}}, the interaction
between both factors is included in the formula.}
}
\description{
Get a string representing a simple linear model two-factor explanation of a
response variable.
}
