% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vector_text_to_numeric}
\alias{vector_text_to_numeric}
\title{Convert string to numeric}
\usage{
vector_text_to_numeric(x)
}
\arguments{
\item{x}{Character.}
}
\description{
Convert a string consisting of numbers seperated by \code{,} to a numeric vector.
}
