% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacDesign.R
\name{FacDesign}
\alias{FacDesign}
\title{FacDesign}
\description{
R6Class representing a factorial design.
}
\section{Usage}{

\preformatted{fac_design <- FacDesign$new(name, k)
}
}

\section{Methods}{

\describe{
  \item{\code{new(name, factor_storage, rep = 1, center = 0, response_name =
  "response", index_name = "index", response = NULL}}{
  Initialize the factorial design.
    \tabular{ll}{
      \code{name} \tab \code{\link[base:character]{Character}}. The factorial
      design's name \cr
      \code{factor_nodes} \cr An object of class \code{\link{ExplorerNode}}
      with children of explorer_class "fac_design_factor_explorer_class"
      storing objects of class \code{\link{FacDesignFactor}}. \cr
      \code{rep} \tab \code{\link[base:numeric]{Numeric}} value giving the
      number of replicates per factor combination. \cr
      \code{center} \tab Currently defunct. \code{\link[base:integer]{Integer}}.
      Number of center points. \cr
      \code{response_name} \tab \code{\link[base:character]{Character}}. Name
      of the response variable. \cr
      \code{index_name} \tab \code{\link[base:character]{Character}}. Name of
      the index variable. \cr
      \code{response} \tab \code{\link[base:numeric]{Numeric}} vector
      containing the values of the response variable for the standardised
      factorial design. If \code{response = NULL}, you have to add a response
      later via \code{this$add_response()}.
    }
  }
  \item{\code{add_response(response)}}{Add or update the response of the
  factorial design.
    \tabular{ll}{
      \code{response} \tab \code{\link[base:numeric]{Numeric}} vector
      containing the values of the response variable for the standardised
      factorial design.
    }
  }
  \item{\code{get_name()}}{Get the factorial design's name.
  }
  \item{\code{get_names()}}{
  Get a list with the following entries:
    \tabular{ll}{
      fac \tab The variable factors' names. \cr
      index \tab The index' name. \cr
      name \tab The factorial design's name. \cr
      response \tab The factorial design response's name.
    }
   The arguments decide wheter a name is part of the list or not.
  }
  \item{\code{get_table(randomized = TRUE, index = TRUE, factors = TRUE,
  response = TRUE, center = FALSE)}}{Get a \code{\link[tibble:tibble]{tibble}}
  representing the factorial design.
    \tabular{ll}{
      \code{randomized} \tab If \code{\link[base:logical]{TRUE}}, the rows are
      ordered in run order and not in standardised order. \cr
      \code{index} \tab If \code{\link[base:logical]{TRUE}}, the index column
      is included in the output. \cr
      \code{factors} \tab If \code{\link[base:logical]{TRUE}}, the factor
      columns are included in the output. \cr
      \code{response} \tab If \code{\link[base:logical]{TRUE}}, the response
      column is included in the output. \cr
      \code{center} \tab If \code{\link[base:logical]{TRUE}}, the rows
      containing observations for the center points are included in the output.
    }
  }
  \item{\code{has_response()}}{Returns a \code{\link[base:logical]{logical}}
  indicating whether the factorial design has a response or not.
  }
  \item{\code{nrow()}}{Get the number of rows of the factorial design.
  }
  \item{\code{rename_fac_names(new_names, old_names = NULL)}}{Set the factors'
  names. If \code{old_names = NULL} the length of \code{new_names} must be
  equal to the number of factors, otherwise the length of \code{new_names}
  and \code{old_names} has to be equal.
    \tabular{ll}{
      \code{new_names} \tab \code{\link[base:character]{Character}}
      vector. \cr
      \code{old_names} \tab \code{\link[base:character]{Character}}
      vector or \code{\link[base:NULL]{NULL}}.
    }
  }
  \item{\code{rename_index_name(new_name)}}{Rename the index  with the
  \code{\link[base:character]{character}} \code{new_name}.
  }
  \item{\code{rename_name(new_name)}}{Rename the factorial design with the
  \code{\link[base:character]{character}} \code{new_name}.
  }
  \item{\code{rename_response_name(new_name)}}{Rename the response with the
  \code{\link[base:character]{character}} \code{new_name}.
  }
}
}

