% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectStorage.R
\name{ObjectStorage}
\alias{ObjectStorage}
\title{ObjectStorage}
\description{
R6Class for storing similar R6 objects.
}
\section{Usage}{

\preformatted{storage = ObjectStorage$new(allowed_classes = NULL)

storage$add_object(object)

storage$get_object(id)
}
}

\section{Methods}{

\describe{
  \item{\code{new(allowed_classes = NULL)}}{Initialize the storage.
    \describe{
      \item{\code{allowed_classes}}{Character vector. \code{class(object)} has to
      return at least one of these classes for being added to the storage.}
    }
  }
  \item{\code{add_object(object)}}{Add an object to the storage.
    \describe{
      \item{\code{object}}{R6 object with public method \code{get_id()}.}
    }
  }
  \item{\code{get_ids()}}{Get the ids of the stored objects as a
    character vector.
  }
  \item{\code{get_object(id, lazy = FALSE)}}{Get an object from the storage
    with \code{object$get_id() == id}.
    \describe{
      \item{\code{id}}{id of an R6 object.}
      \item{\code{lazy}}{If \code{\link[base:logical]{TRUE}}, allow a id
        which is not present in the ids of the storage for a short time.
      }
    }
  }
  \item{\code{get_objects(ids)}}{Get a list of objects from the storage
    with \code{object$get_id() \%in\% ids}.
    \describe{
      \item{\code{ids}}{Character vector. Each element has to be a id of
        an object in the storage.
      }
    }
  }
}
}

