% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectStorage.R
\name{ObjectStorage}
\alias{ObjectStorage}
\title{ObjectStorage}
\description{
R6Class for storing similar R6 objects. These objects all need to implement
the method get_id() and get_name(). The id of an object is
unique in a storage whereas multiple objects can share the same name.
}
\section{Usage}{

\preformatted{storage = ObjectStorage$new(allowed_classes = NULL)

storage$add_object(object)

storage$get_object(id)
}
}

\section{Methods}{

\describe{
  \item{\code{new(allowed_classes = NULL)}}{Initialize the storage.
    \tabular{ll}{
      \code{allowed_classes} \tab Character vector. \code{class(object)} has to
      return at least one of these classes for being added to the storage.
    }
  }
  \item{\code{add_object(object)}}{Add an object to the storage:
    \tabular{ll}{
      \code{object} \tab R6 object with public methods \code{get_id()},
        and \code{get_name()}.
    }
  }
  \item{\code{get_ids()}}{Get the ids of the stored objects as a character
    vector named with the object's names.
  }
  \item{\code{get_length()}}{Get the length of the storage.
  }
  \item{\code{get_names()}}{Get the objects' names, which aren't necessarily
    unique. See also \code{this$get_ids()}.
  }
  \item{\code{get_nth_object(n)}}{Get the object with index \code{n} going
  from \code{1} to \code{this$get_length()}.
  }
  \item{\code{get_object(id)}}{Get an object from the storage
    with \code{object$get_id() == id}.
    \tabular{ll}{
      \code{id} \tab id of an R6 object.
    }
  }
  \item{\code{get_objects(ids)}}{Get a list of objects from the storage
    with \code{object$get_id() \%in\% ids}. If \code{ids} is missing get
    a list with all objects.
    \tabular{ll}{
      \code{ids} \tab Character vector. Each element has to be a id of
        an object in the storage.
    }
  }
}
}

