% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_vector_input.R
\name{observed_vector_input}
\alias{observed_vector_input}
\alias{observed_vector_input_ui}
\title{Observed vector input}
\usage{
observed_vector_input(input, output, session, .data, .values, .parent,
  .label = NULL, .value = "", .error_controller_text = NULL,
  .error_controller_value = NULL)

observed_vector_input_ui(id)
}
\arguments{
\item{input}{Called by \code{\link[shiny]{callModule}}.}

\item{output}{Called by \code{\link[shiny]{callModule}}.}

\item{session}{Called by \code{\link[shiny]{callModule}}.}

\item{.data}{The \code{.data} object.}

\item{.values}{The \code{.values} list.}

\item{.parent}{The parent \code{\link{Node}} object.}

\item{.label}{The contents of the button or link-usually a text label, but
you could also use any other HTML, like an image.}

\item{.value}{Function without an argument (e.g
\code{\link[shiny:reactive]{reactive}}) or vector containing the initial
value of the text input.}

\item{id}{The module's id.}
}
\description{
Observed vector input
}
